refactor: Remove not used code.
